@@ -416,7 +416,6 @@
         "@types/express": "^4.16.4",
         "@types/fs-extra": "^8.1.0",
         "@types/glob": "^7.1.1",
-        "@types/jest": "^25.2.3",
         "@types/md5-file": "^4.0.0",
         "@types/mocha": "^7.0.2",
         "@types/node": "^13.13.42",
@@ -428,21 +427,21 @@
         "@typescript-eslint/eslint-plugin": "^3.10.1",
         "@typescript-eslint/parser": "^3.10.1",
         "copy-webpack-plugin": "^6.4.1",
-        "eslint": "^7.20.0",
+        "eslint": "^7.32.0",
         "eslint-config-prettier": "^6.15.0",
         "eslint-plugin-prettier": "^3.3.1",
         "eslint-plugin-react": "^7.22.0",
         "eslint-plugin-react-hooks": "^4.2.0",
         "glob": "^7.1.6",
         "jest": "^26.6.3",
-        "mocha": "^7.1.1",
+        "mocha": "^8.1.3",
         "prettier": "2.0.5",
         "tmp": "^0.1.0",
         "ts-loader": "^7.0.5",
         "typescript": "^3.9.9",
         "vscode-test": "^1.3.0",
-        "webpack": "^4.46.0",
-        "webpack-cli": "^3.3.12"
+        "webpack": "^5.67.0",
+        "webpack-cli": "^4.9.1"
     },
     "dependencies": {
         "body-parser": "^1.19.0",
@@ -456,6 +455,7 @@
         "react-autosize-textarea": "^7.1.0",
         "react-dom": "^16.14.0",
         "rimraf": "^3.0.2",
+        "sanitize-filename": "^1.6.3",
         "tmp": "^0.1.0"
     }
 }
