@@ -48,11 +48,11 @@ import {
     editorChanged,
     editorClosed,
     checkLaunchWebview,
-} from './webview/editorChange';
+} from "./webview/editorChange";
 
-import { getRetainWebviewContextPref } from './webview/types'
+import { getRetainWebviewContextPref } from "./webview/types";
 
-import JudgeViewProvider from './webview/JudgeView';
+import JudgeViewProvider from "./webview/JudgeView";
 import { RunTestCases } from "./webview/core";
 
 let judgeViewProvider: JudgeViewProvider;
@@ -64,9 +64,9 @@ export const getJudgeViewProvider = () => {
 // Create a new problem
 export async function addProblem() {
     // Use default site when creating a problem from the vscode.
-    let site: SiteDescription = EMPTY;
+    const site: SiteDescription = EMPTY;
 
-    let id = await vscode.window.showInputBox({
+    const id = await vscode.window.showInputBox({
         placeHolder: site.problemIdPlaceholder,
     });
 
@@ -78,7 +78,7 @@ export async function addProblem() {
     let path = getSolutionPath();
     path = join(path!, site.name, "single");
 
-    let problemPath = newProblemFromId(path, site, id);
+    const problemPath = newProblemFromId(path, site, id);
 
     await vscode.commands.executeCommand(
         "vscode.openFolder",
@@ -94,13 +94,13 @@ function parseNumberOfProblems(numberOfProblems: string | undefined) {
     numberOfProblems = numberOfProblems.toLowerCase();
 
     if (numberOfProblems.length === 1) {
-        let value = numberOfProblems.charCodeAt(0);
+        const value = numberOfProblems.charCodeAt(0);
         if (97 <= value && value <= 122) {
             return value - 97 + 1;
         }
     }
 
-    let res = Number.parseInt(numberOfProblems);
+    const res = Number.parseInt(numberOfProblems);
 
     if (Number.isNaN(res)) {
         return undefined;
@@ -110,12 +110,12 @@ function parseNumberOfProblems(numberOfProblems: string | undefined) {
 }
 
 async function addContest() {
-    let path = getSolutionPath();
-    let site: SiteDescription = EMPTY;
+    const path = getSolutionPath();
+    const site: SiteDescription = EMPTY;
 
     let id = undefined;
 
-    let name = await vscode.window.showInputBox({
+    const name = await vscode.window.showInputBox({
         placeHolder: site.contestIdPlaceholder,
     });
 
@@ -124,10 +124,10 @@ async function addContest() {
         return;
     }
 
-    let probCountStr = await vscode.window.showInputBox({
+    const probCountStr = await vscode.window.showInputBox({
         placeHolder: "Number of problems",
     });
-    let probCount = parseNumberOfProblems(probCountStr);
+    const probCount = parseNumberOfProblems(probCountStr);
 
     if (probCount === undefined) {
         vscode.window.showErrorMessage("Number of problems not provided.");
@@ -136,7 +136,7 @@ async function addContest() {
 
     id = name + "-" + probCount.toString();
 
-    let contestPath = await newContestFromId(path!, site, id);
+    const contestPath = await newContestFromId(path!, site, id);
 
     vscode.commands.executeCommand(
         "vscode.openFolder",
@@ -157,10 +157,10 @@ async function debugTestCase(path: string, tcId: string) {
         ],
     });
 
-    let sol = mainSolution(path);
-    let inp = join(path, TESTCASES, `${tcId}.in`);
-    let ans = join(path, TESTCASES, `${tcId}.ans`);
-    let out = join(path, TESTCASES, `${tcId}.out`);
+    const sol = mainSolution(path);
+    const inp = join(path, TESTCASES, `${tcId}.in`);
+    const ans = join(path, TESTCASES, `${tcId}.ans`);
+    const out = join(path, TESTCASES, `${tcId}.out`);
 
     await vscode.commands.executeCommand(
         "vscode.open",
@@ -194,24 +194,24 @@ async function debugTestCase(path: string, tcId: string) {
 }
 
 async function runSolution() {
-    let path_ = currentProblem();
+    const path_ = currentProblem();
 
     if (path_.isNone()) {
         vscode.window.showErrorMessage("No active problem");
         return;
     }
 
-    let path = path_.unwrap();
+    const path = path_.unwrap();
     debug("run-solution", `${path}`);
 
     await vscode.window.activeTextEditor?.document.save().then(() => {
-        let result_ = testSolution(path);
+        const result_ = testSolution(path);
 
         if (result_.isNone()) {
             return;
         }
 
-        let result = result_.unwrap();
+        const result = result_.unwrap();
 
         if (result.isOk()) {
             vscode.window.showInformationMessage(
@@ -220,7 +220,7 @@ async function runSolution() {
         } else if (result.status === Verdict.NO_TESTCASES) {
             vscode.window.showErrorMessage(`No testcases.`);
         } else {
-            let failTestCaseId = result.getFailTestCaseId();
+            const failTestCaseId = result.getFailTestCaseId();
             vscode.window.showErrorMessage(
                 `${verdictName(result.status)} on test ${failTestCaseId}`
             );
@@ -230,19 +230,19 @@ async function runSolution() {
 }
 
 async function compile() {
-    let path_ = currentProblem();
+    const path_ = currentProblem();
 
     if (path_.isNone()) {
         vscode.window.showErrorMessage("No active problem");
         return;
     }
 
-    let path = path_.unwrap();
+    const path = path_.unwrap();
 
     debug("compile", `${path}`);
 
-    let sol = mainSolution(path);
-    let out = join(path, ATTIC, "sol");
+    const sol = mainSolution(path);
+    const out = join(path, ATTIC, "sol");
 
     if (!existsSync(sol)) {
         vscode.window.showErrorMessage("Open a coding environment first.");
@@ -251,7 +251,7 @@ async function compile() {
 
     await vscode.window.activeTextEditor?.document.save().then(() => {
         // Compile solution
-        let result = preRun(sol, out, path!, FRIEND_TIMEOUT);
+        const result = preRun(sol, out, path!, FRIEND_TIMEOUT);
 
         if (result.isNone()) {
             vscode.window.showInformationMessage("No compilation needed.");
@@ -262,21 +262,23 @@ async function compile() {
             } else if (result.unwrap().stderr().length > 0) {
                 message = " with warnings";
             }
-            vscode.window.showInformationMessage("Compilation successful" + message + ".");
+            vscode.window.showInformationMessage(
+                "Compilation successful" + message + "."
+            );
         }
     });
 }
 
 async function openTestCase() {
-    let path_ = currentProblem();
+    const path_ = currentProblem();
 
     if (path_.isNone()) {
         vscode.window.showErrorMessage("No active problem");
         return;
     }
 
-    let path = path_.unwrap();
-    let testCases: any[] = [];
+    const path = path_.unwrap();
+    const testCases: any[] = [];
 
     // Read testcases
     readdirSync(join(path, TESTCASES))
@@ -284,7 +286,7 @@ async function openTestCase() {
             return extname(testCasePath) === ".in";
         })
         .map(function (testCasePath) {
-            let name = removeExtension(testCasePath);
+            const name = removeExtension(testCasePath);
 
             testCases.push({
                 label: name,
@@ -292,13 +294,13 @@ async function openTestCase() {
             });
         });
 
-    let tc = await vscode.window.showQuickPick(testCases, {
+    const tc = await vscode.window.showQuickPick(testCases, {
         placeHolder: "Select test case",
     });
 
     if (tc !== undefined) {
-        let inp = join(path, TESTCASES, `${tc.target}.in`);
-        let out = join(path, TESTCASES, `${tc.target}.ans`);
+        const inp = join(path, TESTCASES, `${tc.target}.in`);
+        const out = join(path, TESTCASES, `${tc.target}.ans`);
 
         await vscode.commands.executeCommand("vscode.setEditorLayout", {
             orientation: 0,
@@ -318,22 +320,22 @@ async function openTestCase() {
 }
 
 async function addTestCase() {
-    let path_ = currentProblem();
+    const path_ = currentProblem();
 
     if (path_.isNone()) {
         vscode.window.showErrorMessage("No active problem");
         return;
     }
 
-    let path = path_.unwrap();
+    const path = path_.unwrap();
 
     let index = 0;
     while (existsSync(join(path, TESTCASES, `${index}.hand.in`))) {
         index += 1;
     }
 
-    let inp = join(path, TESTCASES, `${index}.hand.in`);
-    let out = join(path, TESTCASES, `${index}.hand.ans`);
+    const inp = join(path, TESTCASES, `${index}.hand.in`);
+    const out = join(path, TESTCASES, `${index}.hand.ans`);
 
     writeFileSync(inp, "");
     writeFileSync(out, "");
@@ -359,20 +361,20 @@ async function coding() {
         ter.hide();
     }); //hides terminals
 
-    let path_ = currentProblem();
+    const path_ = currentProblem();
 
     if (path_.isNone()) {
         vscode.window.showErrorMessage("No active problem");
         return;
     }
 
-    let path = path_.unwrap();
+    const path = path_.unwrap();
 
     await vscode.commands.executeCommand("vscode.setEditorLayout", {
         groups: [{}],
     });
 
-    let sol = mainSolution(path);
+    const sol = mainSolution(path);
 
     await vscode.commands.executeCommand(
         "vscode.open",
@@ -382,16 +384,16 @@ async function coding() {
 }
 
 async function stress() {
-    let path_ = currentProblem();
+    const path_ = currentProblem();
 
     if (path_.isNone()) {
         vscode.window.showErrorMessage("No active problem");
         return;
     }
 
-    let path = path_.unwrap();
+    const path = path_.unwrap();
 
-    let _stressTimes: number | undefined = vscode.workspace
+    const _stressTimes: number | undefined = vscode.workspace
         .getConfiguration("acmx.stress", null)
         .get("times");
 
@@ -408,14 +410,14 @@ async function stress() {
 }
 
 async function upgrade() {
-    let path_ = currentProblem();
+    const path_ = currentProblem();
 
     if (path_.isNone()) {
         vscode.window.showErrorMessage("No active problem");
         return;
     }
 
-    let path = path_.unwrap();
+    const path = path_.unwrap();
 
     upgradeArena(path);
 }
@@ -427,25 +429,25 @@ function fileList(dir: string): string[] {
 }
 
 async function setChecker() {
-    let path_ = currentProblem();
+    const path_ = currentProblem();
 
     if (path_.isNone()) {
         vscode.window.showErrorMessage("No active problem");
         return;
     }
 
-    let path = path_.unwrap();
+    const path = path_.unwrap();
 
-    let allCheckersPlain = fileList(join(pathToStatic(), "checkers")).filter(
+    const allCheckersPlain = fileList(join(pathToStatic(), "checkers")).filter(
         (name: string) => name !== "testlib.h"
     );
-    let allChecker = allCheckersPlain.map((value: string) => {
-        var data = readFileSync(
+    const allChecker = allCheckersPlain.map((value: string) => {
+        const data = readFileSync(
             join(pathToStatic(), "checkers", value),
             "utf8"
         );
-        var regex = /setName\("(.*?)"\)/;
-        var matched = regex.exec(data)![1];
+        const regex = /setName\("(.*?)"\)/;
+        const matched = regex.exec(data)![1];
         return {
             label: value.slice(0, value.length - 4),
             detail: matched,
@@ -453,7 +455,7 @@ async function setChecker() {
         };
     });
 
-    let checkerInfo = await vscode.window.showQuickPick(allChecker, {
+    const checkerInfo = await vscode.window.showQuickPick(allChecker, {
         placeHolder: "Select custom checker.",
     });
 
@@ -462,19 +464,19 @@ async function setChecker() {
         return;
     }
 
-    let checker = checkerInfo.target;
+    const checker = checkerInfo.target;
 
-    let checkerPath = join(pathToStatic(), "checkers", checker);
-    let checkerDest = join(path, ATTIC, "checker.cpp");
+    const checkerPath = join(pathToStatic(), "checkers", checker);
+    const checkerDest = join(path, ATTIC, "checker.cpp");
 
     copyFileSync(checkerPath, checkerDest);
-    let config = ConfigFile.loadConfig(path).unwrapOr(ConfigFile.empty());
+    const config = ConfigFile.loadConfig(path).unwrapOr(ConfigFile.empty());
     config.checker = Option.some(checkerDest);
     config.dump(path);
 }
 
 export async function selectDebugTestCase(uriPath: vscode.Uri) {
-    let testCaseName = basename(uriPath.path);
+    const testCaseName = basename(uriPath.path);
     let path = dirname(uriPath.path);
 
     const MAX_DEPTH = 2;
@@ -483,89 +485,89 @@ export async function selectDebugTestCase(uriPath: vscode.Uri) {
         path = dirname(path);
     }
 
-    let launchTaskPath = join(path, ".vscode", "launch.json");
+    const launchTaskPath = join(path, ".vscode", "launch.json");
 
     if (!existsSync(launchTaskPath)) {
         return undefined;
     }
 
-    let launchTaskData = readFileSync(launchTaskPath, "utf8");
+    const launchTaskData = readFileSync(launchTaskPath, "utf8");
     // TODO(#20): Don't use regular expression to replace this. Use JSON parser instead.
-    let newTaskData = launchTaskData.replace(
-        /\"args\"\:.+/,
+    const newTaskData = launchTaskData.replace(
+        /"args":.+/,
         `"args": ["<", "\${fileDirname}/testcases/${testCaseName}"],`
     );
-    let launchTaskFdW = openSync(launchTaskPath, "w");
+    const launchTaskFdW = openSync(launchTaskPath, "w");
     writeSync(launchTaskFdW, newTaskData);
     closeSync(launchTaskFdW);
 }
 
 async function selectMainSolution(uriPath: vscode.Uri) {
-    let path_ = currentProblem();
+    const path_ = currentProblem();
 
     if (path_.isNone()) {
         vscode.window.showErrorMessage("No active problem");
         return;
     }
 
-    let path = path_.unwrap();
+    const path = path_.unwrap();
 
-    let config = ConfigFile.loadConfig(path).unwrapOr(ConfigFile.empty());
+    const config = ConfigFile.loadConfig(path).unwrapOr(ConfigFile.empty());
     config.mainSolution = Option.some(uriPath.path);
     config.dump(path);
     console.log(uriPath);
 }
 
 async function selectBruteSolution(uriPath: vscode.Uri) {
-    let path_ = currentProblem();
+    const path_ = currentProblem();
 
     if (path_.isNone()) {
         vscode.window.showErrorMessage("No active problem");
         return;
     }
 
-    let path = path_.unwrap();
+    const path = path_.unwrap();
 
-    let config = ConfigFile.loadConfig(path).unwrapOr(ConfigFile.empty());
+    const config = ConfigFile.loadConfig(path).unwrapOr(ConfigFile.empty());
     config.bruteSolution = Option.some(uriPath.path);
     config.dump(path);
     console.log(uriPath);
 }
 
 async function selectGenerator(uriPath: vscode.Uri) {
-    let path_ = currentProblem();
+    const path_ = currentProblem();
 
     if (path_.isNone()) {
         vscode.window.showErrorMessage("No active problem");
         return;
     }
 
-    let path = path_.unwrap();
+    const path = path_.unwrap();
 
-    let config = ConfigFile.loadConfig(path).unwrapOr(ConfigFile.empty());
+    const config = ConfigFile.loadConfig(path).unwrapOr(ConfigFile.empty());
     config.generator = Option.some(uriPath.path);
     config.dump(path);
     console.log(uriPath);
 }
 
 async function selectChecker(uriPath: vscode.Uri) {
-    let path_ = currentProblem();
+    const path_ = currentProblem();
 
     if (path_.isNone()) {
         vscode.window.showErrorMessage("No active problem");
         return;
     }
 
-    let path = path_.unwrap();
+    const path = path_.unwrap();
 
-    let config = ConfigFile.loadConfig(path).unwrapOr(ConfigFile.empty());
+    const config = ConfigFile.loadConfig(path).unwrapOr(ConfigFile.empty());
     config.checker = Option.some(uriPath.path);
     config.dump(path);
     console.log(uriPath);
 }
 
-function assignedCopyToClipboardCommand(): Boolean {
-    let generateCodeCommand:
+function assignedCopyToClipboardCommand(): boolean {
+    const generateCodeCommand:
         | string
         | undefined = vscode.workspace
         .getConfiguration("acmx.configuration", null)
@@ -577,26 +579,26 @@ function assignedCopyToClipboardCommand(): Boolean {
 // in code ready for submission. This is useful when we want to apply
 // some filters to the code, or some libraries should be inlined.
 function codeToSubmit(): string | undefined {
-    let path_ = currentProblem();
+    const path_ = currentProblem();
 
     if (path_.isNone()) {
         vscode.window.showErrorMessage("No active problem");
         return;
     }
 
-    let path = path_.unwrap();
-    let generateCodeCommand:
+    const path = path_.unwrap();
+    const generateCodeCommand:
         | string
         | undefined = vscode.workspace
         .getConfiguration("acmx.configuration", null)
         .get("copyToClipboardCommand");
-    let sol = mainSolution(path);
+    const sol = mainSolution(path);
     let content = "";
 
     if (generateCodeCommand === undefined || generateCodeCommand === "") {
         content = readFileSync(sol, "utf8");
     } else {
-        let generateCodeCommands = generateCodeCommand!.split(" ");
+        const generateCodeCommands = generateCodeCommand!.split(" ");
 
         for (let i = 0; i < generateCodeCommands.length; i++) {
             generateCodeCommands[i] = generateCodeCommands[i].replace(
@@ -605,7 +607,7 @@ function codeToSubmit(): string | undefined {
             );
         }
 
-        let execution = runSingle(generateCodeCommands, FRIEND_TIMEOUT, "");
+        const execution = runSingle(generateCodeCommands, FRIEND_TIMEOUT, "");
 
         if (execution.failed()) {
             vscode.window.showErrorMessage("Fail generating submission.");
@@ -619,7 +621,7 @@ function codeToSubmit(): string | undefined {
 }
 
 async function copySubmissionToClipboard() {
-    let content = codeToSubmit();
+    const content = codeToSubmit();
 
     if (content !== undefined) {
         clipboardy.writeSync(content);
@@ -628,16 +630,16 @@ async function copySubmissionToClipboard() {
 }
 
 export async function submitSolution() {
-    let path_ = currentProblem();
+    const path_ = currentProblem();
 
     if (path_.isNone()) {
         vscode.window.showErrorMessage("No active problem");
         return;
     }
 
-    let path = path_.unwrap();
-    let config = ConfigFile.loadConfig(path, true).unwrap();
-    let compileResult = getMainSolutionPath(path, config);
+    const path = path_.unwrap();
+    const config = ConfigFile.loadConfig(path, true).unwrap();
+    const compileResult = getMainSolutionPath(path, config);
 
     if (compileResult.isNone()) {
         vscode.window.showErrorMessage("Could not get the code");
@@ -647,7 +649,7 @@ export async function submitSolution() {
     let mainSolutionPath = compileResult.unwrap().code;
     debug("submit-main-solution-path", `${mainSolutionPath}`);
 
-    let url_ = config.url();
+    const url_ = config.url();
 
     if (url_.isNone()) {
         vscode.window.showErrorMessage("No active url");
@@ -666,10 +668,10 @@ export async function submitSolution() {
     }
 
     if (assignedCopyToClipboardCommand()) {
-        let content = codeToSubmit();
-        let outputFile = tmp.fileSync();
+        const content = codeToSubmit();
+        const outputFile = tmp.fileSync();
         writeSync(outputFile.fd, content);
-        let nameWithExtension =
+        const nameWithExtension =
             outputFile.name + getExtension(mainSolutionPath);
         renameSync(outputFile.name, nameWithExtension);
         mainSolutionPath = nameWithExtension;
@@ -682,21 +684,21 @@ export async function submitSolution() {
     debug("submit-solution-command", `${submitCommand}`);
 
     await vscode.window.activeTextEditor?.document.save().then(() => {
-        let ter = acmxTerminal();
+        const ter = acmxTerminal();
         ter.show();
         ter.sendText(submitCommand!);
     });
 }
 
 async function editLanguage() {
-    let languages: any[] = [];
+    const languages: any[] = [];
 
     readdirSync(globalLanguagePath())
         .filter(function (testCasePath) {
             return extname(testCasePath) === ".json";
         })
         .map(function (testCasePath) {
-            let name = removeExtension(testCasePath);
+            const name = removeExtension(testCasePath);
 
             languages.push({
                 label: name,
@@ -704,7 +706,7 @@ async function editLanguage() {
             });
         });
 
-    let selectedLanguage = await vscode.window.showQuickPick(languages, {
+    const selectedLanguage = await vscode.window.showQuickPick(languages, {
         placeHolder: "Select language",
     });
 
@@ -726,80 +728,86 @@ export function activate(context: vscode.ExtensionContext) {
     initAcmX();
     startCompetitiveCompanionService();
 
-    let addProblemCommand = vscode.commands.registerCommand(
+    const addProblemCommand = vscode.commands.registerCommand(
         "acmx.addProblem",
         addProblem
     );
-    let addContestCommand = vscode.commands.registerCommand(
+    const addContestCommand = vscode.commands.registerCommand(
         "acmx.addContest",
         addContest
     );
-    let runSolutionCommand = vscode.commands.registerCommand(
+    const runSolutionCommand = vscode.commands.registerCommand(
         "acmx.runSolution",
         runSolution
     );
-    let openTestCaseCommand = vscode.commands.registerCommand(
+    const openTestCaseCommand = vscode.commands.registerCommand(
         "acmx.openTestCase",
         openTestCase
     );
-    let addTestCaseCommand = vscode.commands.registerCommand(
+    const addTestCaseCommand = vscode.commands.registerCommand(
         "acmx.addTestCase",
         addTestCase
     );
-    let codingCommand = vscode.commands.registerCommand("acmx.coding", coding);
-    let stressCommand = vscode.commands.registerCommand("acmx.stress", stress);
-    let upgradeCommand = vscode.commands.registerCommand(
+    const codingCommand = vscode.commands.registerCommand(
+        "acmx.coding",
+        coding
+    );
+    const stressCommand = vscode.commands.registerCommand(
+        "acmx.stress",
+        stress
+    );
+    const upgradeCommand = vscode.commands.registerCommand(
         "acmx.upgrade",
         upgrade
     );
-    let compileCommand = vscode.commands.registerCommand(
+    const compileCommand = vscode.commands.registerCommand(
         "acmx.compile",
         compile
     );
-    let setCheckerCommand = vscode.commands.registerCommand(
+    const setCheckerCommand = vscode.commands.registerCommand(
         "acmx.setChecker",
         setChecker
     );
-    let selectDebugTestCaseCommand = vscode.commands.registerCommand(
+    const selectDebugTestCaseCommand = vscode.commands.registerCommand(
         "acmx.selectDebugTestCase",
         selectDebugTestCase
     );
-    let selectMainSolutionCommand = vscode.commands.registerCommand(
+    const selectMainSolutionCommand = vscode.commands.registerCommand(
         "acmx.selectMainSolution",
         selectMainSolution
     );
-    let selectBruteSolutionCommand = vscode.commands.registerCommand(
+    const selectBruteSolutionCommand = vscode.commands.registerCommand(
         "acmx.selectBruteSolution",
         selectBruteSolution
     );
-    let selectGeneratorCommand = vscode.commands.registerCommand(
+    const selectGeneratorCommand = vscode.commands.registerCommand(
         "acmx.selectGenerator",
         selectGenerator
     );
-    let selectCheckerCommand = vscode.commands.registerCommand(
+    const selectCheckerCommand = vscode.commands.registerCommand(
         "acmx.selectChecker",
         selectChecker
     );
-    let copySubmissionToClipboardCommand = vscode.commands.registerCommand(
+    const copySubmissionToClipboardCommand = vscode.commands.registerCommand(
         "acmx.copyToClipboard",
         copySubmissionToClipboard
     );
-    let editLanguageCommand = vscode.commands.registerCommand(
+    const editLanguageCommand = vscode.commands.registerCommand(
         "acmx.editLanguage",
         editLanguage
     );
 
-    let submitSolutionCommand = vscode.commands.registerCommand(
+    const submitSolutionCommand = vscode.commands.registerCommand(
         "acmx.submitSolution",
         submitSolution
     );
 
-    let debugTestCommand = vscode.commands.registerCommand(
+    const debugTestCommand = vscode.commands.registerCommand(
         "acmx.debugTest",
         debugTest
     );
 
-    let runTestCasesCommand = vscode.commands.registerCommand(
+    const runTestCasesCommand = vscode.commands.registerCommand(
         "acmx.runTestCases",
         RunTestCases
     );
@@ -834,7 +842,7 @@ export function activate(context: vscode.ExtensionContext) {
             webviewOptions: {
                 retainContextWhenHidden: getRetainWebviewContextPref(),
             },
-        },
+        }
     );
 
     context.subscriptions.push(webviewView);
@@ -852,12 +860,9 @@ export function activate(context: vscode.ExtensionContext) {
     vscode.window.onDidChangeVisibleTextEditors((editors) => {
         if (editors.length === 0) {
             getJudgeViewProvider().extensionToJudgeViewMessage({
-                command: 'new-problem',
+                command: "new-problem",
                 problem: undefined,
             });
         }
     });
 }
-
-// this method is called when your extension is deactivated
-export function deactivate() {}
