@@ -1,10 +1,7 @@
-import * as vscode from 'vscode';
-import { VSToWebViewMessage, WebviewToVSEvent } from './types';
-import runAllAndSave from './processRunAll';
-import {
-    getAutoShowJudgePref,
-    getRetainWebviewContextPref,
-} from './types';
+import * as vscode from "vscode";
+import { VSToWebViewMessage, WebviewToVSEvent } from "./types";
+import runAllAndSave from "./processRunAll";
+import { getAutoShowJudgePref, getRetainWebviewContextPref } from "./types";
 
 import {
     getProblemForDocument,
@@ -14,12 +11,12 @@ import {
     deleteProblemFile,
     deleteProblemCase,
     setOnlineJudgeEnv,
-    selectProblemCase
-} from './core';
-import { runSingleAndSave } from './processRunSingle';
+    selectProblemCase,
+} from "./core";
+import { runSingleAndSave } from "./processRunSingle";
 
 class JudgeViewProvider implements vscode.WebviewViewProvider {
-    public static readonly viewType = 'cph.judgeView';
+    public static readonly viewType = "cph.judgeView";
 
     private _view?: vscode.WebviewView;
 
@@ -48,91 +45,94 @@ class JudgeViewProvider implements vscode.WebviewViewProvider {
 
         webviewView.webview.onDidReceiveMessage(
             async (message: WebviewToVSEvent) => {
-                console.log('Got from webview', message);
+                console.log("Got from webview", message);
                 switch (message.command) {
-                    case 'run-single-and-save': {
+                    case "run-single-and-save": {
                         const problem = message.problem;
                         const id = message.id;
                         runSingleAndSave(problem, id);
                         break;
                     }
 
-                    case 'run-all-and-save': {
+                    case "run-all-and-save": {
                         const problem = message.problem;
                         runAllAndSave(problem);
                         break;
                     }
 
-                    case 'save': {
+                    case "save": {
                         saveProblem(message.problem.srcPath, message.problem);
                         break;
                     }
 
-                    case 'kill-running': {
+                    case "kill-running": {
                         // killRunning();
                         break;
                     }
 
-                    case 'select-case': {
+                    case "select-case": {
                         selectProblemCase(message.problem.srcPath, message.id);
                         break;
                     }
 
-                    case 'delete-tcs': {
+                    case "delete-tcs": {
                         if (message.id) {
-                            deleteProblemCase(message.problem.srcPath, message.id);
+                            deleteProblemCase(
+                                message.problem.srcPath,
+                                message.id
+                            );
                             break;
                         }
                         this.extensionToJudgeViewMessage({
-                            command: 'new-problem',
+                            command: "new-problem",
                             problem: undefined,
                         });
                         deleteProblemFile(message.problem.srcPath);
                         break;
                     }
 
-                    case 'submitCf': {
+                    case "submitCf": {
                         SubmitProblem(message.problem);
                         break;
                     }
-                    case 'submitKattis': {
+                    case "submitKattis": {
                         // submitKattisProblem(message.problem);
                         break;
                     }
 
-                    case 'online-judge-env': {
+                    case "online-judge-env": {
                         setOnlineJudgeEnv(message.value);
                         break;
                     }
 
-                    case 'get-initial-problem': {
+                    case "get-initial-problem": {
                         this.getInitialProblem();
                         break;
                     }
 
-                    case 'create-local-problem': {
+                    case "create-local-problem": {
                         AddProblem();
                         break;
                     }
 
                     default: {
-                        console.error('Unknown event received from webview');
+                        console.error("Unknown event received from webview");
                     }
                 }
-            },
+            }
         );
     }
 
     private getInitialProblem() {
         const doc = vscode.window.activeTextEditor?.document;
         this.extensionToJudgeViewMessage({
-            command: 'new-problem',
+            command: "new-problem",
             problem: getProblemForDocument(doc),
         });
 
         // also load any messages from before that were lost.
         this.messageBuffer.forEach((message) => {
-            console.log('Restored buffer command', message.command);
+            console.log("Restored buffer command", message.command);
             this._view?.webview.postMessage(message);
         });
 
@@ -144,9 +144,9 @@ class JudgeViewProvider implements vscode.WebviewViewProvider {
     public problemPath: string | undefined;
 
     public async focus() {
-        console.log('focusing');
+        console.log("focusing");
         if (!this._view) {
-            await vscode.commands.executeCommand('cph.judgeView.focus');
+            await vscode.commands.executeCommand("cph.judgeView.focus");
         } else {
             this._view.show?.(true);
         }
@@ -156,15 +156,15 @@ class JudgeViewProvider implements vscode.WebviewViewProvider {
         console.log(message.command);
 
         switch (message.command) {
-            case 'waiting-for-submit':
-            case 'compiling-start':
-            case 'run-all': {
+            case "waiting-for-submit":
+            case "compiling-start":
+            case "run-all": {
                 this.focus();
             }
         }
 
         if (
-            message.command === 'new-problem' &&
+            message.command === "new-problem" &&
             message.problem !== undefined &&
             getAutoShowJudgePref()
         ) {
@@ -174,7 +174,7 @@ class JudgeViewProvider implements vscode.WebviewViewProvider {
 
     /** Posts a message to the webview. */
     public extensionToJudgeViewMessage = async (
-        message: VSToWebViewMessage,
+        message: VSToWebViewMessage
     ) => {
         this.focusIfNeeded(message);
         if (
@@ -184,10 +184,10 @@ class JudgeViewProvider implements vscode.WebviewViewProvider {
             // Always focus on the view whenever a command is posted. Meh.
             // this._view.show?.(true); // `show` is not implemented in 1.49 but is for 1.50 insiders
             this._view.webview.postMessage(message);
-            if (message.command !== 'submit-finished') {
-                console.log('View got message', message);
+            if (message.command !== "submit-finished") {
+                console.log("View got message", message);
             }
-            if (message.command === 'new-problem') {
+            if (message.command === "new-problem") {
                 if (message.problem === undefined) {
                     this.problemPath = undefined;
                 } else {
@@ -195,8 +195,8 @@ class JudgeViewProvider implements vscode.WebviewViewProvider {
                 }
             }
         } else {
-            if (message.command !== 'new-problem') {
-                console.log('Pushing to buffer', message.command);
+            if (message.command !== "new-problem") {
+                console.log("Pushing to buffer", message.command);
                 this.messageBuffer.push(message);
             } else {
                 this.messageBuffer = [];
@@ -206,15 +206,15 @@ class JudgeViewProvider implements vscode.WebviewViewProvider {
 
     private _getHtmlForWebview(webview: vscode.Webview) {
         const css = webview.asWebviewUri(
-            vscode.Uri.joinPath(this._extensionUri, 'dist', 'app.css'),
+            vscode.Uri.joinPath(this._extensionUri, "dist", "app.css")
         );
 
         const js = webview.asWebviewUri(
             vscode.Uri.joinPath(
                 this._extensionUri,
-                'dist',
-                'frontend.module.js',
-            ),
+                "dist",
+                "frontend.module.js"
+            )
         );
 
         const html = `
